backwards compatible

+++ silo-core/test/foundry/Silo/getters/RawLiquidityAndProtectedCollateral.i.sol
@@ -101,7 +101,7 @@ contract RawLiquidityAndProtectedCollateralTest is SiloLittleHelper, Test {
 
         (uint256 collateralAssets, uint256 protectedAssets) = _silo.getCollateralAndProtectedTotalsStorage();
         uint256 debtAssets = _silo.getDebtAssets();
-        (uint160 daoAndDeployerRevenue,,,,,) = _silo.getSiloStorage();
+        (uint192 daoAndDeployerRevenue,,,,) = _silo.getSiloStorage();
         uint256 liquidity = _silo.getRawLiquidity();
 
         emit log_named_uint("collateralAssets", collateralAssets);
