+++ silo-core/test/foundry/utils/hook-receivers/liquidation/LiquidationCall_1token.i.sol
@@ -132,8 +132,8 @@ contract LiquidationCall1TokenTest is SiloLittleHelper, Test {
 
         ISiloConfig.ConfigData memory debtConfig = siloConfig.getConfig(address(silo0));
 
-        (, uint32 interestRateTimestamp0,,,,) = silo0.getSiloStorage();
-        (, uint32 interestRateTimestamp1,,,,) = silo1.getSiloStorage();
+        (, uint64 interestRateTimestamp0,,,) = silo0.getSiloStorage();
+        (, uint64 interestRateTimestamp1,,,) = silo1.getSiloStorage();
         assertEq(interestRateTimestamp0, 1, "interestRateTimestamp0 is 1 because we deposited and borrow same asset");
         assertEq(block.timestamp, 1, "block.timestamp");
 
@@ -243,8 +243,8 @@ contract LiquidationCall1TokenTest is SiloLittleHelper, Test {
         }
 
         { // too deep
-            (, uint32 interestRateTimestamp0After,,,,) = silo0.getSiloStorage();
-            (, uint32 interestRateTimestamp1After,,,,) = silo1.getSiloStorage();
+            (, uint64 interestRateTimestamp0After,,,) = silo0.getSiloStorage();
+            (, uint64 interestRateTimestamp1After,,,) = silo1.getSiloStorage();
 
             assertEq(interestRateTimestamp0 + timeForward, interestRateTimestamp0After, "interestRateTimestamp #0");
             assertGt(interestRateTimestamp1After, 0, "interestRateTimestamp #1 (because of withdraw)");
@@ -343,8 +343,8 @@ contract LiquidationCall1TokenTest is SiloLittleHelper, Test {
 
         ISiloConfig.ConfigData memory debtConfig = siloConfig.getConfig(address(silo1));
 
-        (, uint32 interestRateTimestamp0,,,,) = silo0.getSiloStorage();
-        (, uint32 interestRateTimestamp1,,,,) = silo1.getSiloStorage();
+        (, uint64 interestRateTimestamp0,,,) = silo0.getSiloStorage();
+        (, uint64 interestRateTimestamp1,,,) = silo1.getSiloStorage();
 
         // move forward with time so we can have interests
 
@@ -446,8 +446,8 @@ contract LiquidationCall1TokenTest is SiloLittleHelper, Test {
         { // too deep
             uint256 timeForward = 120 days; // MUST MATCH previous declaration, this copy is because coverage too deep
 
-            (, uint32 interestRateTimestamp0After,,,,) = silo0.getSiloStorage();
-            (, uint32 interestRateTimestamp1After,,,,) = silo1.getSiloStorage();
+            (, uint64 interestRateTimestamp0After,,,) = silo0.getSiloStorage();
+            (, uint64 interestRateTimestamp1After,,,) = silo1.getSiloStorage();
 
             assertEq(interestRateTimestamp0 + timeForward, interestRateTimestamp0After, "interestRateTimestamp #0");
 
@@ -472,8 +472,8 @@ contract LiquidationCall1TokenTest is SiloLittleHelper, Test {
 
         ISiloConfig.ConfigData memory debtConfig = siloConfig.getConfig(address(silo0));
 
-        (, uint32 interestRateTimestamp0,,,,) = silo0.getSiloStorage();
-        (, uint32 interestRateTimestamp1,,,,) = silo1.getSiloStorage();
+        (, uint64 interestRateTimestamp0,,,) = silo0.getSiloStorage();
+        (, uint64 interestRateTimestamp1,,,) = silo1.getSiloStorage();
 
         // move forward with time so we can have interests
 
@@ -570,8 +570,8 @@ contract LiquidationCall1TokenTest is SiloLittleHelper, Test {
         { // too deep
             uint256 timeForward = 150 days; // MUST MATCH previous, this copy is because coverage throws too deep
 
-            (, uint32 interestRateTimestamp0After,,,,) = silo0.getSiloStorage();
-            (, uint32 interestRateTimestamp1After,,,,) = silo1.getSiloStorage();
+            (, uint64 interestRateTimestamp0After,,,) = silo0.getSiloStorage();
+            (, uint64 interestRateTimestamp1After,,,) = silo1.getSiloStorage();
 
             assertEq(interestRateTimestamp0 + timeForward, interestRateTimestamp0After, "interestRateTimestamp #0");
             assertLt(interestRateTimestamp1, interestRateTimestamp1After, "interestRateTimestamp #1");
