+++ silo-core/test/foundry/utils/hook-receivers/liquidation/LiquidationCall_2tokens.i.sol
@@ -146,8 +146,8 @@ contract LiquidationCall2TokensTest is SiloLittleHelper, Test {
         ISiloConfig.ConfigData memory collateralConfig = siloConfig.getConfig(address(silo1));
         ISiloConfig.ConfigData memory debtConfig = siloConfig.getConfig(address(silo0));
 
-        (, uint32 interestRateTimestamp0,,,,) = silo0.getSiloStorage();
-        (, uint32 interestRateTimestamp1,,,,) = silo1.getSiloStorage();
+        (, uint64 interestRateTimestamp0,,,) = silo0.getSiloStorage();
+        (, uint64 interestRateTimestamp1,,,) = silo1.getSiloStorage();
 
         (
             uint256 collateralToLiquidate, uint256 debtToRepay, bool sTokenRequired
@@ -230,8 +230,8 @@ contract LiquidationCall2TokensTest is SiloLittleHelper, Test {
         }
 
         { // too deep
-            (, uint32 interestRateTimestamp0After,,,,) = silo0.getSiloStorage();
-            (, uint32 interestRateTimestamp1After,,,,) = silo1.getSiloStorage();
+            (, uint64 interestRateTimestamp0After,,,) = silo0.getSiloStorage();
+            (, uint64 interestRateTimestamp1After,,,) = silo1.getSiloStorage();
 
             assertEq(interestRateTimestamp0 + timeForward, interestRateTimestamp0After, "interestRateTimestamp #0");
             assertEq(interestRateTimestamp1 + timeForward, interestRateTimestamp1After, "interestRateTimestamp #1");
@@ -324,8 +324,8 @@ contract LiquidationCall2TokensTest is SiloLittleHelper, Test {
         ISiloConfig.ConfigData memory collateralConfig = siloConfig.getConfig(address(silo0));
         ISiloConfig.ConfigData memory debtConfig = siloConfig.getConfig(address(silo1));
 
-        (, uint32 interestRateTimestamp0,,,,) = silo0.getSiloStorage();
-        (, uint32 interestRateTimestamp1,,,,) = silo1.getSiloStorage();
+        (, uint64 interestRateTimestamp0,,,) = silo0.getSiloStorage();
+        (, uint64 interestRateTimestamp1,,,) = silo1.getSiloStorage();
 
         // move forward with time so we can have interests
 
@@ -400,8 +400,8 @@ contract LiquidationCall2TokensTest is SiloLittleHelper, Test {
         */
 
         { // too deep
-            (, uint32 interestRateTimestamp0After,,,,) = silo0.getSiloStorage();
-            (, uint32 interestRateTimestamp1After,,,,) = silo1.getSiloStorage();
+            (, uint64 interestRateTimestamp0After,,,) = silo0.getSiloStorage();
+            (, uint64 interestRateTimestamp1After,,,) = silo1.getSiloStorage();
 
             assertEq(interestRateTimestamp0 + timeForward, interestRateTimestamp0After, "interestRateTimestamp #0");
             assertEq(interestRateTimestamp1 + timeForward, interestRateTimestamp1After, "interestRateTimestamp #1");
@@ -564,7 +564,7 @@ contract LiquidationCall2TokensTest is SiloLittleHelper, Test {
             uint256 collateralToLiquidate, uint256 debtToRepay,
         ) = partialLiquidation.maxLiquidation(BORROWER);
 
-        (uint160 daoAndDeployerRevenue,,,,,) = silo1.getSiloStorage();
+        (uint192 daoAndDeployerRevenue,,,,) = silo1.getSiloStorage();
         uint256 maxRepay = silo1.maxRepay(BORROWER);
         uint256 interest = maxRepay - DEBT - daoAndDeployerRevenue / 1e18;
         uint256 liquidity = silo1.getLiquidity();
